Character physics updated to higher percision; various macros added.
The macros would be cool, if it worked. (Should've commited more...)
The problem right now is that the macros provided by RGBDS
are too limited to work this way.
Uncertain whether to drop this or to seek an alternative assembler.

Added title graphics
joypad.z80: Added streak for A button
memory.z80: Added a lot of macros.
    Struct traversal works great, Label definitions don't.
objects.z80:
    Updated to use struct traversal macros.
    Added jumping to player input.
sprites.z80: Updated to use new macros. Broken.
variables.z80:
    Updated character struct to higher percision.
    Added a charNext field to make struct traversal macros nicer to use.
    Added jump button streak var.