Rewrite TickObject_Physics to actually work right
Movement doesn't work right yet, but Roxy now moves.
(These characters needed names. For no reason I'm calling the snow fox
Roxy and the gecko Leo. Those feel like fitting names?)

The old code didn't read/write to the correct offsets in the char struct
which meant that movement could not happen, even when the dX or dY
members were forcibly set to non-zero values.
Also, work has been done to simplify the acceletation code (akthough
it's very broken right now) and to convert it to use the new named
register macros.

Elsewhere, a bug has been fixed in GetScreenX was fixed so that it
loaded FieldX with the proper endinaness, which made the math below make
sense again. Same issue in incFieldX, except now it also increments
instead of decrementing. Silly me. :P

Next steps is to ensure that the update code works right (this hasn't
been heavily tested) and then to read up on how the heck acceleration
works in an old game like this because I'm lost.